requirements.txt
Django==1.11.13
MySQL-python==1.2.5
Pillow==5.0.0
Pygments==2.2.0
django-detect==1.0.5
django-ip==1.0.2
django-json-response==1.1.5
django-markdown-deux==1.0.5
django-mobi2==1.0.2
isort==4.3.4
marrow.mailer==4.0.2
markdown2==2.3.5
pngquant==1.0.6
pycodestyle=2.4.0
pytz==2018.4
qiniu==7.2.2
requests==2.19.1
shortuuid==0.5.0
xpinyin==0.5.5
